Reorganise inotify code
