+++ src/config/setup/config_setup_bool.cc
@@ -28,8 +28,11 @@
 #include "config/config_options.h"
 
 #ifdef HAVE_INOTIFY
-#include "util/mt_inotify.h"
+#include "content/inotify/mt_inotify.h"
 #endif
+
+#define B_TRUE "true"
+#define B_FALSE "false"
 
 void ConfigBoolSetup::makeOption(const pugi::xml_node& root, const std::shared_ptr<Config>& config, const std::map<std::string, std::string>* arguments)
 {
@@ -39,19 +42,19 @@ void ConfigBoolSetup::makeOption(const pugi::xml_node& root, const std::shared_p
 
 static bool validateTrueFalse(const std::string& optValue)
 {
-    return (optValue == "true" || optValue == "false");
+    return (optValue == B_TRUE || optValue == B_FALSE);
 }
 
 static bool validateYesNo(std::string_view value)
 {
-    return value == "yes" || value == "no";
+    return value == YES || value == NO;
 }
 
 void ConfigBoolSetup::makeOption(std::string optValue, const std::shared_ptr<Config>& config, const std::map<std::string, std::string>* arguments)
 {
     if (!validateTrueFalse(optValue) && !validateYesNo(optValue))
         throw_std_runtime_error("Invalid {} value {}", xpath, optValue);
-    optionValue = std::make_shared<BoolOption>(optValue == YES || optValue == "true");
+    optionValue = std::make_shared<BoolOption>(optValue == YES || optValue == B_TRUE);
     setOption(config);
 }
 
@@ -69,7 +72,7 @@ bool ConfigBoolSetup::checkValue(std::string& optValue, const std::string& pathN
 
     if (!validateTrueFalse(optValue) && !validateYesNo(optValue))
         throw_std_runtime_error("Invalid {}/{} value {}", pathName, xpath, optValue);
-    return optValue == YES || optValue == "true";
+    return optValue == YES || optValue == B_TRUE;
 }
 
 std::shared_ptr<ConfigOption> ConfigBoolSetup::newOption(bool optValue)
@@ -78,29 +81,31 @@ std::shared_ptr<ConfigOption> ConfigBoolSetup::newOption(bool optValue)
     return optionValue;
 }
 
+#define IN_AUTO "auto"
+
 bool ConfigBoolSetup::CheckInotifyValue(std::string& value)
 {
     bool tempBool = false;
-    if ((value != "auto") && !validateYesNo(value)) {
+    if ((value != IN_AUTO) && !validateYesNo(value) && !validateTrueFalse(value)) {
         log_error("Error in config file: incorrect parameter for \"<autoscan use-inotify=\" attribute");
         return false;
     }
 
 #ifdef HAVE_INOTIFY
     bool inotifySupported = Inotify::supported();
-    tempBool = (inotifySupported && value == "auto");
+    tempBool = (inotifySupported && value == IN_AUTO);
 #endif
 
-    if (value == YES) {
+    if (value == YES || value == B_TRUE) {
 #ifdef HAVE_INOTIFY
         if (!inotifySupported) {
-            log_error("You specified \"yes\" in \"<autoscan use-inotify=\"\">"
+            log_error("You specified \"" YES "\" in \"<autoscan use-inotify=\"\">"
                       " however your system does not have inotify support");
             return false;
         }
         tempBool = true;
 #else
-        log_error("You specified \"yes\" in \"<autoscan use-inotify=\"\">"
+        log_error("You specified \"" YES "\" in \"<autoscan use-inotify=\"\">"
                   " however this version of Gerbera was compiled without inotify support");
         return false;
 #endif
