+++ src/content/inotify/mt_inotify.h
@@ -36,8 +36,7 @@
 
 #ifdef HAVE_INOTIFY
 
-#include <sys/inotify.h>
-
+#include "inotify_types.h"
 #include "util/grb_fs.h"
 
 /// \brief Inotify interface.
@@ -53,7 +52,7 @@ public:
     /// \param path file or directory to monitor.
     /// \param events inotify event mask
     /// \return watch descriptor or a negative value on error
-    int addWatch(const fs::path& path, uint32_t events, unsigned int retryCount) const;
+    int addWatch(const fs::path& path, InotifyFlags events, unsigned int retryCount) const;
 
     /// \brief Removes a previously added file or directory from the watch list
     /// \param wd watch descriptor that was returned by the add_watch function
